refactor(looker): separate out `lkml/` and `api/` interfaces
