+++ python_modules/libraries/dagster-looker/dagster_looker_tests/lkml/test_asset_specs.py
@@ -3,8 +3,11 @@ from typing import Any, Mapping, Optional, Sequence, Tuple
 
 import pytest
 from dagster import AssetKey, Definitions
-from dagster_looker.asset_specs import build_looker_asset_specs
-from dagster_looker.dagster_looker_translator import DagsterLookerTranslator, LookMLStructureType
+from dagster_looker.lkml.asset_specs import build_looker_asset_specs
+from dagster_looker.lkml.dagster_looker_translator import (
+    DagsterLookerTranslator,
+    LookMLStructureType,
+)
 
 from dagster_looker_tests.looker_projects import (
     test_exception_derived_table_path,
